Move tests from datagen to casemapping crate
